requirements.txt
streamlit>=1.31
pandas>=2.0
pillow>=10.0
